Restore support for virtual serial devices on OSX (#10).
This commit restores support for virtual serial devices by making the
logic implemented in #7 optional. The rationale for #7 was to allow for
non-standard baud-rates by using an OSX-specific IOCTL that only works
on real serial devices.

This change tests for the baud-rate and decides upon a strategy: if the
baud-rate is standard, we use the regular, portable IOCTL which works on
all devices. If the baud-rate isn't standard, we use the OSX-specific
IOCTL introduced in #7.